Move screen recording logic from Room component into a separate component called Record
